+++ src/pymetropolis/metro_demand/population/common.py
@@ -1,0 +1,3 @@
+# List of activity purposes in the survey data.
+# Note. For now these are MobiSurvStd purposes. Ideally it should be configurable.
+PURPOSES = ["home", "work", "education", "shopping", "task", "leisure", "escort", "other"]